refactor(import): rename import packages to avoid Java keyword conflicts
