+++ app/src/main/kotlin/com/kuvaszuptime/kuvasz/services/monitor/MonitorImporter.kt
@@ -3,15 +3,15 @@ package com.kuvaszuptime.kuvasz.services.monitor
 import com.kuvaszuptime.kuvasz.config.HttpMonitorConfig
 import com.kuvaszuptime.kuvasz.config.IcmpMonitorConfig
 import com.kuvaszuptime.kuvasz.config.PushMonitorConfig
-import com.kuvaszuptime.kuvasz.models.dto.import.MonitorImportResultDto
+import com.kuvaszuptime.kuvasz.models.dto.importing.MonitorImportResultDto
 import com.kuvaszuptime.kuvasz.models.monitor.http.toMonitorRecord
 import com.kuvaszuptime.kuvasz.models.monitor.icmp.toMonitorRecord
 import com.kuvaszuptime.kuvasz.models.monitor.push.toMonitorRecord
 import com.kuvaszuptime.kuvasz.repositories.HttpMonitorRepository
 import com.kuvaszuptime.kuvasz.repositories.IcmpMonitorRepository
 import com.kuvaszuptime.kuvasz.repositories.PushMonitorRepository
-import com.kuvaszuptime.kuvasz.services.monitor.import.ImportStrategy
-import com.kuvaszuptime.kuvasz.services.monitor.import.ValidatedMonitorImport
+import com.kuvaszuptime.kuvasz.services.monitor.importer.ImportStrategy
+import com.kuvaszuptime.kuvasz.services.monitor.importer.ValidatedMonitorImport
 import com.kuvaszuptime.kuvasz.util.transactionResultWithError
 import com.kuvaszuptime.kuvasz.validation.IntegrationIdValidator
 import jakarta.inject.Singleton
