--- app/src/main/kotlin/com/kuvaszuptime/kuvasz/services/monitor/import/MonitorImportParser.kt
@@ -1,7 +1,0 @@
-package com.kuvaszuptime.kuvasz.services.monitor.import
-
-import com.kuvaszuptime.kuvasz.models.dto.import.MonitorImportDto
-
-interface MonitorImportParser {
-    fun parse(content: ByteArray): MonitorImportDto
-}
