--- app/src/main/kotlin/com/kuvaszuptime/kuvasz/services/monitor/import/MonitorImportService.kt
@@ -1,8 +1,0 @@
-package com.kuvaszuptime.kuvasz.services.monitor.import
-
-import com.kuvaszuptime.kuvasz.models.dto.import.MonitorImportDto
-import com.kuvaszuptime.kuvasz.models.dto.import.MonitorImportResultDto
-
-interface MonitorImportService {
-    fun importMonitors(importDto: MonitorImportDto, dryRun: Boolean = false): MonitorImportResultDto
-}
