+++ app/src/main/kotlin/com/kuvaszuptime/kuvasz/services/monitor/importer/DefaultMonitorImportService.kt
@@ -1,10 +1,10 @@
-package com.kuvaszuptime.kuvasz.services.monitor.import
+package com.kuvaszuptime.kuvasz.services.monitor.importer
 
-import com.kuvaszuptime.kuvasz.models.dto.import.HttpMonitorImportAdapter
-import com.kuvaszuptime.kuvasz.models.dto.import.IcmpMonitorImportAdapter
-import com.kuvaszuptime.kuvasz.models.dto.import.MonitorImportDto
-import com.kuvaszuptime.kuvasz.models.dto.import.MonitorImportResultDto
-import com.kuvaszuptime.kuvasz.models.dto.import.PushMonitorImportAdapter
+import com.kuvaszuptime.kuvasz.models.dto.importing.HttpMonitorImportAdapter
+import com.kuvaszuptime.kuvasz.models.dto.importing.IcmpMonitorImportAdapter
+import com.kuvaszuptime.kuvasz.models.dto.importing.MonitorImportDto
+import com.kuvaszuptime.kuvasz.models.dto.importing.MonitorImportResultDto
+import com.kuvaszuptime.kuvasz.models.dto.importing.PushMonitorImportAdapter
 import com.kuvaszuptime.kuvasz.services.monitor.MonitorImporter
 import jakarta.inject.Singleton
 import jakarta.validation.ValidationException
