+++ app/src/main/kotlin/com/kuvaszuptime/kuvasz/services/monitor/importer/ImportStrategy.kt
@@ -1,6 +1,6 @@
-package com.kuvaszuptime.kuvasz.services.monitor.import
+package com.kuvaszuptime.kuvasz.services.monitor.importer
 
-import com.kuvaszuptime.kuvasz.models.dto.import.MonitorTypeImportResult
+import com.kuvaszuptime.kuvasz.models.dto.importing.MonitorTypeImportResult
 import org.jooq.DSLContext
 
 interface ImportStrategy {
