+++ app/src/main/kotlin/com/kuvaszuptime/kuvasz/services/monitor/importer/MonitorImportParser.kt
@@ -1,0 +1,7 @@
+package com.kuvaszuptime.kuvasz.services.monitor.importer
+
+import com.kuvaszuptime.kuvasz.models.dto.importing.MonitorImportDto
+
+interface MonitorImportParser {
+    fun parse(content: ByteArray): MonitorImportDto
+}
