+++ app/src/main/kotlin/com/kuvaszuptime/kuvasz/services/monitor/importer/MonitorImportService.kt
@@ -1,0 +1,8 @@
+package com.kuvaszuptime.kuvasz.services.monitor.importer
+
+import com.kuvaszuptime.kuvasz.models.dto.importing.MonitorImportDto
+import com.kuvaszuptime.kuvasz.models.dto.importing.MonitorImportResultDto
+
+interface MonitorImportService {
+    fun importMonitors(importDto: MonitorImportDto, dryRun: Boolean = false): MonitorImportResultDto
+}
