+++ model/src/main/kotlin/com/kuvaszuptime/kuvasz/models/dto/importing/MonitorImportResultDto.kt
@@ -1,4 +1,4 @@
-package com.kuvaszuptime.kuvasz.models.dto.import
+package com.kuvaszuptime.kuvasz.models.dto.importing
 
 import com.kuvaszuptime.kuvasz.models.MonitorType
 import io.micronaut.core.annotation.Introspected
